tpl: Add workaround for Go template data race
Closes #5926
